+++ projects/qgrid-core/edit/edit.row.shortcut.factory.js
@@ -1,7 +1,7 @@
 export function editRowShortcutFactory(plugin) {
     const { model } = plugin;
     return action => {
-        const shortcuts = model.edit()[action + 'Shortcuts'];
+        const shortcuts = model.edit()[action + 'Shortcut'];
         return shortcuts && (shortcuts['row'] || shortcuts['$default']) ;
     };
 }